Move view command into own file
